requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
pyyaml
dvc
dvc-azure
dvc-s3
pytest
mlflow
numpy